changed location of LoginPage to Pages/, changed styling to be more responsive
