+++ client/src/components/InputTextField/InputTextField.css
@@ -5,8 +5,8 @@
   border: 2px solid var(--white);
   border-radius: 3px;
   padding: 0.7rem 2rem;
+  margin: 1rem 5rem 1rem 5rem;
   width: 100%;
-  min-width: 320px;
   max-width: 400px;
   font-size: 20px;
   font-family: var(--default-font);
