--- client/src/components/LoginPage/LoginPage.css
@@ -1,41 +1,0 @@
-
-
-.LoginPageForm {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: center;
-    width: 100vw;
-    height: 100vh;
-    padding-top: 65px;
-    background: var(--yellow-gradient);
-    background-size: cover;
-    box-sizing: border-box;
-}
-
-.LogoLogin {
-    height: 240px;
-    width: 240px;
-    box-sizing: border-box;
-}
-
-.InputTextField__container--default {
-    margin: 15px;
-    width: 320px;
-    height: 50px;
-    box-sizing: border-box;
-}
-
-#button {
-    width: 200px;
-    height: 50px;
-    box-sizing: border-box;
-    margin-top: 10px;
-}
-
-.ForgotPass {
-    text-align: center;
-    margin: 0;
-    font-family: var(--default-font);
-    box-sizing: border-box;
-}
